==> notebooks/0.01-mas-dir-and-raw-data-prep.ipynb <==
import sys

sys.path.append('/Workspace/Users/m.bakr@berkeley.edu/261-Final-Project/flightdelays/')
from utils import lst_files_r
from dataset import testfunc


import os
print(os.listdir('/Workspace/Users/m.bakr@berkeley.edu/261-Final-Project/flightdelays'))

data_BASE_DIR = "dbfs:/mnt/mids-w261/"
display(dbutils.fs.ls(f"{data_BASE_DIR}/datasets_final_project_2022/"))

team_BASE_DIR = f"dbfs:/student-groups/Group_4_1"
display(dbutils.fs.ls(f"{team_BASE_DIR}"))

# creating folders as per the checkpointing strategy (No need to rerun again)
dbutils.fs.mkdirs(f"{team_BASE_DIR}/raw")
dbutils.fs.mkdirs(f"{team_BASE_DIR}/interim")
dbutils.fs.mkdirs(f"{team_BASE_DIR}/processed")
dbutils.fs.mkdirs(f"{team_BASE_DIR}/external") # for external data not in the original data folder

# Refresh raw data to the raw folders reading from the source (don't rerun unless there is an updated raw datasets)

# flight data
dbutils.fs.cp(f"{data_BASE_DIR}/datasets_final_project_2022/parquet_airlines_data_3m/", f"{team_BASE_DIR}/raw/flightdelays/parquet_airlines_data_3m", recurse=True)
dbutils.fs.cp(f"{data_BASE_DIR}/datasets_final_project_2022/parquet_airlines_data_1y/", f"{team_BASE_DIR}/raw/flightdelays/parquet_airlines_data_1y", recurse=True)

# Weather data
dbutils.fs.cp(f"{data_BASE_DIR}/datasets_final_project_2022/parquet_weather_data_3m/", f"{team_BASE_DIR}/raw/weather/parquet_weather_data_3m", recurse=True)

# Airport data (stations)
dbutils.fs.cp(f"{data_BASE_DIR}/datasets_final_project_2022/stations_data/", f"{team_BASE_DIR}/raw/stations", recurse=True)

# OTPW data
qdf_otpw = spark.read.format("csv").option("header","true").load(f"dbfs:/mnt/mids-w261/OTPW_3M_2015.csv")
qdf_otpw.write.mode("overwrite").parquet(f"{team_BASE_DIR}/raw/OTPW/OTPW_3M_2015.parquet")

# Airport codes Table (no need to rerun unless there is an updated raw datasets)
url = "https://raw.githubusercontent.com/datasets/airport-codes/refs/heads/main/data/airport-codes.csv"
import pandas as pd
airport_codes = pd.read_csv(url)

# saving as parquet file for future use
spark.createDataFrame(airport_codes).write.mode("overwrite").parquet(f"{team_BASE_DIR}/external/airport_codes.parquet/")
