==> notebooks/0.08-mas-data-join-pipeline.ipynb <==
!pip install timezonefinder

from pyspark.sql import functions as F
from pyspark.sql.window import Window
from pyspark.sql.types import StringType, DoubleType
from datetime import datetime, timedelta
from timezonefinder import TimezoneFinder
import pytz

## Variables and Directories

# Variables and directories
data_BASE_DIR = "dbfs:/mnt/mids-w261/datasets_final_project_2022"
team_BASE_DIR = f"dbfs:/student-groups/Group_4_1"
spark.sparkContext.setCheckpointDir(f"{team_BASE_DIR}/checkpoints")
period = "" # one of the following values ("", "_3m", "_6m", "_1y")

# Datasets
flights = spark.read.parquet(f"{data_BASE_DIR}/parquet_airlines_data{period}")
# cleaned weather data from "https://dbc-fae72cab-cf59.cloud.databricks.com/editor/notebooks/1017866335725443?o=4021782157704243#command/6573382849454181"
weather = spark.read.parquet(f"{team_BASE_DIR}/interim/weather_{period}_checkpoint")
airport_codes = spark.read.parquet(f"{team_BASE_DIR}/external/airport_codes.parquet/")
# stations = spark.read.parquet(f"{data_BASE_DIR}/stations_data/stations_with_neighbors.parquet/")
stations = spark.read.parquet(f"{team_BASE_DIR}/interim/stations_data/stations_with_neighbors.parquet/") # updated
timezones = spark.read.parquet(f"{team_BASE_DIR}/external/tz_lookup_v2.parquet")

print(timezones.count())

display(dbutils.fs.ls(f"{team_BASE_DIR}/interim/weather__checkpoint"))

## Step 1: Cleaning and Feature selection

# Flights Selected Features
f_feat = [
    'QUARTER',
    'MONTH',
    'DAY_OF_MONTH',
    'DAY_OF_WEEK',
    'FL_DATE',
    'OP_UNIQUE_CARRIER',
    'TAIL_NUM',
    'OP_CARRIER_FL_NUM',
    'ORIGIN',
    'DEST',
    'CRS_DEP_TIME',
    'DEP_TIME',
    'DEP_DELAY',
    'TAXI_OUT',
    'TAXI_IN',
    'CRS_ARR_TIME',
    'ARR_TIME',
    'ARR_DELAY',
    'CANCELLED',
    'DIVERTED',
    'CRS_ELAPSED_TIME',
    'ACTUAL_ELAPSED_TIME',
    'AIR_TIME',
    'DISTANCE',
    'YEAR',
    'ORIGIN_CITY_NAME',
    'DEST_CITY_NAME',
    'ORIGIN_AIRPORT_SEQ_ID', # TODO: Check with team agin about this
    'DEST_AIRPORT_SEQ_ID', # TODO: Check with team agin about this
    ]

# Weather Selected Features
w_feat = [
    "DATE",
    "STATION",
    "LATITUDE", 
    "LONGITUDE",
    'NAME',
    "HourlyDewPointTemperature",
    "HourlyDryBulbTemperature",
    "HourlyPrecipitation",
    "HourlyPresentWeatherType",
    "HourlyPressureChange", #?? alot of nulls
    "HourlyRelativeHumidity",
    "HourlySkyConditions", # to be removed
    "HourlyVisibility",
    "HourlyWetBulbTemperature",
    "HourlyWindDirection", # to be removed
    "HourlyWindGustSpeed", #?? alot of nulls
    "HourlyWindSpeed",
    "Sunrise",  #?? alot of nulls
    "Sunset",  #?? alot of nulls
    "REM",
    "weather_datetime_utc"]

# Read Airline Data (flights) selected features and removing duplicates (if any)
flights_clean = flights.select(f_feat).dropDuplicates()
flights_clean.cache()
flights_clean = flights_clean.checkpoint() # Checkpoint after cleaning (prevents recomputation) TODO: save to DBFS  iterim folder)

display(flights.count())
display(flights_clean.count())

## Step 2: Joining Flight with Airports to retrieve ICAO

### 1. Checking missing IATA codes in Airpot codes table and fix or adding the missing

# Extract distinct ORIGIN and DEST codes
flight_iata = flights_clean.select("ORIGIN").union(flights_clean.select("DEST")).distinct()

# Extract distinct IATA codes from airport_codes
airport_iata = airport_codes.select("iata_code", "icao_code").distinct()

# Find missing IATA codes in flights but not in airport data
missing_iata = (
    flight_iata
    .join(airport_iata, flight_iata["ORIGIN"] == airport_iata["iata_code"], "left")
    .filter(
        airport_iata["iata_code"].isNull() | 
        airport_iata["icao_code"].isNull()
        )
    .select(F.col("ORIGIN").alias("missing_code"), "iata_code", "icao_code")
)

display(missing_iata)

[Sloulin Field International Airport IATA code change from ISN to XWA](https://en.wikipedia.org/wiki/Williston_Basin_International_Airport)

September 2015. p. 2. Retrieved October 15, 2016. The current airport (Sloulin Field International Airport – ISN) and proposed replacement airport (Williston Basin International Airport – XWA) will be owned and operated by the City of Williston.
FAA Airport Form 5010 for XWA PDF, effective July 13, 2023.

[McKinney National Airport](https://en.wikipedia.org/wiki/McKinney_National_Airport)

McKinney National Airport (ICAO: KTKI, FAA LID: TKI), formerly Collin County Regional Airport at McKinney

# Fixing missing iata code: (Replacing the ISN to XWA)
flights_clean = flights_clean.na.replace(['ISN'], ['XWA'], 'ORIGIN').na.replace(['ISN'], ['XWA'], 'DEST')

# Add icao code for McKinney National Airport (iata code = TKI, ICAO = KTKI)
airport_codes = (
    airport_codes
    .withColumn(
        "icao_code",
        F.when(F.col("iata_code") == "TKI", "KTKI")
        .otherwise(F.col("icao_code"))
        )
)

### 2. Performing the join to get ICAO codes

# Register as temporary views for SQL use
flights_clean.createOrReplaceTempView("flights_clean")
airport_codes.createOrReplaceTempView("airport_codes")

# Apply Broadcast Join for small airport_codes table
airport_codes_broadcast = F.broadcast(airport_codes)

flights_w_icao = flights_clean \
    .join(airport_codes_broadcast.alias("a1"), flights_clean.ORIGIN == F.col("a1.iata_code"), "left") \
    .join(airport_codes_broadcast.alias("a2"), flights_clean.DEST == F.col("a2.iata_code"), "left") \
    .select(
        flights_clean["*"],
        F.col("a1.icao_code").alias("ORIGIN_ICAO"),
        F.col("a1.type").alias("origin_type"),
        F.col("a1.iso_region").alias("origin_region"),
        F.col("a2.icao_code").alias("DEST_ICAO"),
        F.col("a2.type").alias("dest_type"),
        F.col("a2.iso_region").alias("dest_region")
    )

# Cache & checkpoint post join (TODO: save to DBFS iterim folder)
flights_w_icao.cache()
flights_w_icao = flights_w_icao.checkpoint()

### 3. Checking if flights still have any missing ICAO code after the join

# Get statistics for validation: check for null ORIGIN_ICAO and DEST_ICAO after the join (must be null)

origin_iata_match_count = flights_w_icao.filter(F.col("ORIGIN_ICAO").isNotNull()).count()
dest_iata_match_count = flights_w_icao.filter(F.col("DEST_ICAO").isNotNull()).count()
total_flights_w_icao = flights_w_icao.count()

print(f"Origin IATA match rate: {origin_iata_match_count/total_flights_w_icao:.2%} - {total_flights_w_icao - origin_iata_match_count} IATA codes unmatched")
print(f"Destination IATA match rate: {dest_iata_match_count/total_flights_w_icao:.2%} - {total_flights_w_icao - dest_iata_match_count} IATA codes unmatched")

## Step 3: Joining Flight with Stations to get the closest station ID and coordinates

### 1. Calculate Haversine distance to nearest neighbor

# ============================
# UDF: Haversine Distance
# ============================

from math import radians, cos, sin, atan2, sqrt

def haversine(lat1, lon1, lat2, lon2, unit="km"):
    """
    Calculate the distance between two points on Earth using the Haversine formula.

    Args:
        lon1: Longitude of the first point in decimal degrees.
        lat1: Latitude of the first point in decimal degrees.
        lon2: Longitude of the second point in decimal degrees.
        lat2: Latitude of the second point in decimal degrees.
        units: The units of the distance to return. Can be "km" (kilometers) or "mi" (miles). Default is "km".

    Returns:
        The distance between the two points in kilometers.
    """
    # Radius of the Earth in kilometers 6371 in miles 3956
    if  unit == "km":
        R = 6371
    else:
        R = 3956

    # Convert decimal degrees to radians
    lon1, lat1, lon2, lat2 = map(radians, [lon1, lat1, lon2, lat2])

    # Haversine formula
    dlon = lon2 - lon1
    dlat = lat2 - lat1
    a = sin(dlat/2)**2 + cos(lat1) * cos(lat2) * sin(dlon/2)**2
    c = 2 * atan2(sqrt(a), sqrt(1 - a))
    distance = R * c

    return distance

# Register as UDF
haversine_udf = udf(haversine, DoubleType())

unit = "km"
if f"distance_to_neighbor_{unit}" not in stations.columns:
    # Calculate Haversine distance to nearest neighbor
    stations = stations.withColumn(f"distance_to_neighbor_{unit}", 
                    haversine_udf(F.col("lat"), F.col("lon"), F.col("neighbor_lat"), F.col("neighbor_lon")))

    stations.write.mode("overwrite").parquet(f"{team_BASE_DIR}/interim/stations_data/stations_with_neighbors.parquet")

### 2. Validate Missing ICAO Codes in Stations Data before the join

# rereading satitions table just in case it was saved in the previous step
stations = spark.read.parquet(f"{team_BASE_DIR}/interim/stations_data/stations_with_neighbors.parquet/")

# Extract unique ICAO codes from flights
flight_icao = flights_w_icao.select("ORIGIN_ICAO").union(flights_w_icao.select("DEST_ICAO")).distinct()

# Extract unique station ICAO codes
stations_icao = stations.select("neighbor_call").distinct()

# Find missing ICAO codes
missing_icao_lst = (
    flight_icao.join(stations_icao, flight_icao["ORIGIN_ICAO"] == stations_icao["neighbor_call"], "left_anti")
    .select(F.col("ORIGIN_ICAO").alias("missing_code"))
    .rdd.flatMap(lambda x: x).collect()
)

### 3. Augment Stations with the missing airports and Find Nearest Weather Stations (Optimized with Haversine)

if len(missing_icao_lst) > 0:
    # Construct Missing ICAO Data Using Airport Codes
    missing_icao = (
        airport_codes.filter(F.col("icao_code").isin(missing_icao_lst))
        .withColumn("neighbor_id", F.col("ident"))
        .withColumn("neighbor_name", F.col("name"))
        .withColumn("neighbor_state", F.split(F.col("iso_region"), "-")[1].cast("string"))
        .withColumn("neighbor_call", F.col("icao_code"))
        .withColumn("neighbor_lat", F.split(F.col("coordinates"), ",")[0].cast("double"))
        .withColumn("neighbor_lon", F.split(F.col("coordinates"), ",")[1].cast("double"))
        .select("neighbor_id", "neighbor_name", "neighbor_state", "neighbor_call", "neighbor_lat", "neighbor_lon")
    )

    # Calculate distances between missing ICAO and weather stations
    dist_stations = stations.select("usaf", "wban", "station_id", "lat", "lon").distinct()
    missing_distances = (
        dist_stations.crossJoin(missing_icao)
        .withColumn("distance_to_neighbor", F.lit(""))
        .withColumn(f"distance_to_neighbor_{unit}", haversine_udf(F.col("lat"), F.col("lon"), F.col("neighbor_lat"), F.col("neighbor_lon")))
    )

    aug_stations = stations.union(missing_distances)

    aug_stations.write.mode("overwrite").parquet(f"{team_BASE_DIR}/interim/stations_data/stations_with_neighbors.parquet")
    
else:
    aug_stations = stations

# Find closest station
aug_stations = (
    aug_stations.withColumn("rn", F.row_number().over(Window.partitionBy("neighbor_call").orderBy(f"distance_to_neighbor_{unit}")))
    .filter(F.col("rn") == 1)
)

# Cache & checkpoint augmented stations (TODO: save to DBFS iterim folder)
aug_stations.cache()
aug_stations = aug_stations.checkpoint()

aug_stations.count()

### 4. Performing the join to get the nearest station's data

# Register as temporary views for SQL use
flights_w_icao.createOrReplaceTempView("flights_w_icao")
aug_stations.createOrReplaceTempView("aug_stations")

# Apply Broadcast Join for small aug_stations table
aug_stations_broadcast = F.broadcast(aug_stations)

flights_stations = flights_w_icao \
    .join(aug_stations_broadcast.alias("a1"), flights_w_icao.ORIGIN_ICAO == F.col("a1.neighbor_call"), "left") \
    .join(aug_stations_broadcast.alias("a2"), flights_w_icao.DEST_ICAO == F.col("a2.neighbor_call"), "left") \
    .select(
        flights_w_icao["*"],
        # Origin Station
        F.col("a1.neighbor_name").alias("origin_airport_name"),
        F.col("a1.station_id").alias("origin_station_id"),
        F.col("a1.lat").alias("origin_station_lat"),
        F.col("a1.lon").alias("origin_station_lon"),
        F.col("a1.neighbor_lat").alias("origin_airport_lat"),
        F.col("a1.neighbor_lon").alias("origin_airport_lon"),
        F.col(f"a1.distance_to_neighbor_{unit}").alias("origin_station_dis"),

        # Destination Station
        F.col("a2.neighbor_name").alias("dest_airport_name"),
        F.col("a2.station_id").alias("dest_station_id"),
        F.col("a2.lat").alias("dest_station_lat"),
        F.col("a2.lon").alias("dest_station_lon"),
        F.col("a2.neighbor_lat").alias("dest_airport_lat"),
        F.col("a2.neighbor_lon").alias("dest_airport_lon"),
        F.col(f"a2.distance_to_neighbor_{unit}").alias("dest_station_dis")
    )

# Cache & checkpoint post join (TODO: save to DBFS iterim folder)
flights_stations.cache()
flights_stations = flights_stations.checkpoint()

### 5. Checking if we have any missing stations after the join

# Get statistics for validation: check for null origin_airport_name and dest_airport_name after the join (must be null)

origin_icao_match_count = flights_stations.filter(F.col("origin_airport_lat").isNotNull()).count()
dest_icao_match_count = flights_stations.filter(F.col("dest_airport_lat").isNotNull()).count()
total_stations = flights_stations.count()

print(f"Origin ICAO match rate: {origin_icao_match_count/total_stations:.2%} - {total_stations - origin_icao_match_count} ICAO codes unmatched")
print(f"Destination ICAO match rate: {dest_icao_match_count/total_stations:.2%} - {total_stations - dest_icao_match_count} ICAO codes unmatched")

## Step 4: Handling time

### 1. Validate Missing ICAO in Timezones Data before the join

# Extract unique station ICAO codes
tz_icao = timezones.select("neighbor_call").distinct()

# Find missing ICAO codes in timezones table
missing_tz_lst = (
    flight_icao.join(tz_icao, flight_icao["ORIGIN_ICAO"] == tz_icao["neighbor_call"], "left_anti")
    .select(F.col("ORIGIN_ICAO").alias("missing_code"))
    .rdd.flatMap(lambda x: x).collect()
)

### 2. Finding missing timezones using coordinates

# ============================
# UDF: Timezones Lookup
# ============================
def find_timezone(lat, lng):
    tf = TimezoneFinder()
    timezone_str = tf.timezone_at(lat=lat, lng=lng)
    return timezone_str if timezone_str else "Unknown"


if len(missing_tz_lst) > 0:
    missing_tz = (
            flights_stations.filter(F.col("ORIGIN_ICAO").isin(missing_tz_lst))
            .select(F.col("ORIGIN_ICAO").alias("neighbor_call"),
                    F.col("origin_airport_lat").alias("neighbor_lat"),
                    F.col("origin_airport_lon").alias("neighbor_lon"))
            .union(flights_stations.filter(F.col("DEST_ICAO").isin(missing_tz_lst))
                   .select(F.col("DEST_ICAO").alias("neighbor_call"),
                           F.col("dest_airport_lat").alias("neighbor_lat"),
                           F.col("dest_airport_lon").alias("neighbor_lon"))).distinct())

    # define udf for time zone lookup
    find_timezone_udf = udf(find_timezone, StringType())

    # add time zone column
    missing_tz = missing_tz.withColumn("timezone", find_timezone_udf(F.col("neighbor_lat"), F.col("neighbor_lon")))
    
    # Augmenting Timezones data with the missing airport data
    timezones = timezones.union(missing_tz)
    # re-save timezones data as a parquet file (TODO: uncomment later once verified)
    timezones.write.mode("overwrite").parquet(f"{team_BASE_DIR}/external/tz_lookup_v2.parquet")

### 3. Finding Timezone using time zones helper table

# Register as temporary views for SQL use
flights_stations.createOrReplaceTempView("flights_stations")
timezones.createOrReplaceTempView("timezones")

# Apply Broadcast Join for small timezones table
tz_broadcast = F.broadcast(timezones)

flights_tz = flights_stations \
    .join(tz_broadcast.alias("a1"), flights_stations.ORIGIN_ICAO == F.col("a1.neighbor_call"), "left") \
    .join(tz_broadcast.alias("a2"), flights_stations.DEST_ICAO == F.col("a2.neighbor_call"), "left") \
    .select(
        flights_stations["*"],
        # Origin Station
        F.col("a1.timezone").alias("origin_timezone"),

        # Destination Station
        F.col("a2.timezone").alias("dest_timezone"),
    )

# Cache the data to avoid recomputing the time zones (TODO: save to DBFS iterim folder)
flights_tz.cache()
flights_tz = flights_tz.checkpoint()

### 4. Checking if we have any missing stations after the join

# Get statistics for validation: check for null origin_timezone and dest_timezone after the join (must be null)

origin_tz_match_count = flights_tz.filter(F.col("origin_timezone").isNotNull()).count()
dest_tz_match_count = flights_tz.filter(F.col("dest_timezone").isNotNull()).count()
total_flights_tz = flights_tz.count()

print(f"Origin timeszones match rate: {origin_tz_match_count/total_flights_tz:.2%} - {total_flights_tz - origin_tz_match_count} timeszones unmatched")
print(f"Destination timeszones match rate: {dest_tz_match_count/total_flights_tz:.2%} - {total_flights_tz - dest_tz_match_count} timeszones unmatched")

### 5. Joining Date and time fields

# Joining Derparture Date and Departure time to datetime
flights_tz = flights_tz.withColumn(
    "sched_depart_date_time",
    F.concat_ws(
        "T",
        # Extract just the date part (handles both date and timestamp formats)
        F.date_format(F.to_date(F.col("FL_DATE")), "yyyy-MM-dd"),
        F.date_format(F.to_timestamp(F.concat(F.lpad("CRS_DEP_TIME", 4, "0"), F.lit("00")), "HHmmss"), "HH:mm:ss")
    )
)

### 6. Converting to UTC and adding UTC-2 and UTC-4

# ============================
# UDF: Convert Deprature Time to UTC
# ============================

def to_utc(dt: str, tz: str) -> str:
    if dt is None:
        return None
    dt_format = "%Y-%m-%dT%H:%M:%S"
    local_dt = datetime.strptime(dt, dt_format)
    if tz:
        timezone = pytz.timezone(tz)
        local_dt = timezone.localize(local_dt)
        
        # Convert to UTC
        utc_dt = local_dt.astimezone(pytz.utc)
        return utc_dt.strftime(dt_format)
    return None

utc_udf = F.udf(to_utc)

flights_utc = flights_tz.withColumn("sched_depart_utc", utc_udf(F.col("sched_depart_date_time"), F.col("origin_timezone").cast(F.StringType())))

# Cache the data to avoid recomputing the UTC (TODO: save to DBFS iterim folder)
flights_utc.cache()
flights_utc = flights_utc.checkpoint()

# ============================
# UDF: Calculate Prior Times
# ============================
def calculate_prior_times(utc_datetime_str, hours_prior):
    if utc_datetime_str is None:
        return None
    try:
        # Parse UTC datetime string
        utc_format = "%Y-%m-%dT%H:%M:%S"
        utc_dt = datetime.strptime(utc_datetime_str, utc_format)
        
        # Subtract hours to get prior times
        prior_dt = utc_dt - timedelta(hours=hours_prior)
        return prior_dt.strftime("%Y-%m-%dT%H:%M:%S")
    
    except Exception as e:
        return None

# Register UDFs
calculate_prior_times_udf = F.udf(calculate_prior_times, StringType())

flights_utc_2_4 = flights_utc.withColumn("two_hours_prior_depart_UTC", calculate_prior_times_udf(F.col("sched_depart_utc"), F.lit(2)))
flights_utc_2_4 = flights_utc_2_4.withColumn("four_hours_prior_depart_UTC", calculate_prior_times_udf(F.col("sched_depart_utc"), F.lit(4)))

# Cache the data to avoid recomputing the UTC-2 and UTC-4 (TODO: save to DBFS iterim folder)
flights_utc_2_4.cache()
flights_utc_2_4 = flights_utc_2_4.checkpoint()

## Step 5: Joining Weather Data

### 1. Optimize Flights Data

# Add date columns for partitioning
flights_prep = (flights_utc_2_4
    # Convert timestamps and extract date parts
    .withColumn("sched_depart_utc", F.col("sched_depart_utc").cast("timestamp"))
    .withColumn("two_hours_prior_depart_UTC", F.col("two_hours_prior_depart_UTC").cast("timestamp"))
    .withColumn("four_hours_prior_depart_UTC", F.col("four_hours_prior_depart_UTC").cast("timestamp"))
    # Add date columns for efficient partitioning
    .withColumn("origin_join_date", F.to_date("four_hours_prior_depart_UTC"))
    # Repartition to optimize joins (Helps with shuffle efficiency)
    .repartition("origin_station_id", "sched_depart_utc")
    # Persist for multiple joins
    .persist()
)

### 2. Optimize Weather Data

# Add date column and filter only to needed fields to reduce shuffle size
weather_prep = (weather
                # Select only columns we need for the join and prediction
                .select(w_feat)
                .withColumn("weather_date", F.to_date("weather_datetime_utc"))
                .withColumn("weather_datetime_utc", F.col("weather_datetime_utc").cast("timestamp"))
                # Cluster by date for optimization
                .repartition("weather_date", "STATION")
                .persist()
            )

# Create a broadcast lookup of dates AND stations to filter weather data
flight_date_stations = (flights_prep
    .select(
        F.col("origin_join_date").alias("date"),
        F.col("origin_station_id").alias("station")
    )
    .distinct()
    .persist()
)

# Filter weather data to only include dates AND stations in our flight dataset
weather_filtered = (weather_prep
                    .join(
                        F.broadcast(flight_date_stations),
                        (weather_prep.weather_date == flight_date_stations.date) & 
                        (weather_prep.STATION == flight_date_stations.station),
                        "inner"
                    )
                    .select(weather_prep["*"])
                    .persist()
                )
flight_date_stations.cache()
weather_filtered.cache()

print(weather.count())
print(len(weather.columns))
print(weather_filtered.count())
print(len(weather_filtered.columns))

### 3. Origin Weather Join

# Merge flight and weather data for origin airports
origin_weather = (
    flights_prep.alias("f")
    .join(
        weather_filtered.alias("w"),
        (F.col("f.origin_station_id") == F.col("w.STATION")) &
        (F.col("w.weather_datetime_utc") >= F.col("f.four_hours_prior_depart_UTC")) &
        (F.col("w.weather_datetime_utc") <= F.col("f.two_hours_prior_depart_UTC")),
        "left"
    )
)

origin_prefix = "origin_"
origin_weather_col = [F.col(f"w.{c}").alias(f"{origin_prefix}{c}") for c in weather_filtered.columns]
# Select closest weather reading to 2 hours prior to departure for origin
origin_weather_closest = (
    origin_weather
    .withColumn(
        "time_diff", 
        F.abs(F.col("w.weather_datetime_utc") - F.col("f.two_hours_prior_depart_UTC"))
    )
    .withColumn(
        "rn", 
        F.row_number().over(
            Window
            .partitionBy("OP_CARRIER_FL_NUM", "ORIGIN", "DEST", "TAIL_NUM", "sched_depart_utc")
            .orderBy("time_diff")
        )
    )
    .filter(F.col("rn") == 1)
    .drop("rn", "time_diff")
    # Rename columns to indicate origin weather
    .select(
        "f.*",
        *origin_weather_col
    )
    .persist()
)

origin_weather_closest.cache()
origin_weather_closest = origin_weather_closest.checkpoint()

# Checkpoint the final joined dataframe
origin_weather_closest.write.mode("overwrite").parquet(f"{team_BASE_DIR}/interim/join_checkpoints/joined_flights_weather{period}_v1.parquet")

# Check the final dataframe dimensions
num_rows = origin_weather_closest.count()
num_cols = len(origin_weather_closest.columns)
print(f"DataFrame dimensions: Rows = {num_rows}, Columns = {num_cols}")

### 5. Collect Join Metrics

# Get statistics for validation
origin_match_count = origin_weather_closest.filter(F.col("origin_STATION").isNotNull()).count()
# dest_match_count = dest_weather_closest.filter(F.col("dest_STATION").isNotNull()).count()
total_flights = flights_prep.count()

print(f"Origin weather match rate: {origin_match_count/total_flights:.2%}- {total_flights - origin_match_count} flights unmatched")
# print(f"Destination weather match rate: {dest_match_count/total_flights:.2%} - {total_flights - dest_match_count} flights unmatched")

### 6. Sving and getting the size and dimension of the final joined data

# Checkpoint the final joined dataframe
dest_weather_closest.write.mode("overwrite").parquet(f"{team_BASE_DIR}/interim/join_checkpoints/joined_flights_weather{period}_v1.parquet")

# Calculate the final file size
path=f"{team_BASE_DIR}/interim/join_checkpoints/joined_flights_weather{period}_v1.parquet"
filelist=dbutils.fs.ls(path)
filelist_df = spark.createDataFrame(filelist)
filelist_df.createOrReplaceTempView("adlsSize")

final_flights_size = spark.sql("select round(sum(size)/(1024*1024*1024), 2) as sizeInGB from adlsSize")
display(final_flights_size)

# unpersisting data frames
flights_prep.unpersist()
weather_prep.unpersist()
flight_date_stations.unpersist()
weather_filtered.unpersist()
origin_weather_closest.unpersist()
dest_weather_closest.unpersist()
