==> notebooks/3.14-eil-5y-modeling-SANDBOX.ipynb <==
# Baseline Modeling on Joined Data

Erica Landreth

## Imports and Setup

# Configure Spark settings for better performance
from pyspark.sql import SparkSession
spark = SparkSession.builder\
    .config("spark.executor.memory", "16g")\
    .config("spark.executor.cores", "32g")\
    .config("spark.executor.cores", 4)\
    .appName('Final Project Training')\
    .getOrCreate()
spark.conf.set("spark.sql.shuffle.partitions", "200")
spark.conf.set("spark.default.parallelism", "200")

# pip install prophet

# imports
import pandas as pd
import numpy as np
import pytz
from datetime import datetime, timedelta, time
from prophet import Prophet
from prophet.make_holidays import make_holidays_df
import matplotlib.pyplot as plt
import plotly.express as px
import plotly.graph_objects as go
from pyspark.sql.functions import to_timestamp
from prophet.plot import plot_forecast_component
from pyspark.sql.types import StructType, StructField, StringType, IntegerType, StructType, DoubleType, LongType
from pyspark.ml.feature import VectorAssembler, StandardScaler, StringIndexer, OneHotEncoder, MinMaxScaler
from pyspark.ml.classification import LogisticRegression
from pyspark.mllib.evaluation import MulticlassMetrics,BinaryClassificationMetrics
from pyspark.ml.evaluation import RegressionEvaluator
from pyspark.sql import functions as f
from pyspark.sql.window import Window
from pyspark.sql.functions import col, when, to_timestamp, lit, udf
from pyspark.ml import Pipeline
import seaborn as sns
import matplotlib.pyplot as plt
from pyspark.sql.functions import col, to_timestamp, to_date, when
from prophet.make_holidays import make_holidays_df
from xgboost.spark import SparkXGBClassifier

### Set options

# data time period
period = "" # on of the following values ("", "3m", "6m", "1y")

# number of cross-validation folds and overlap
k = 5
overlap = 0.2

# compute seasonality?
# (False if you've already saved out seasonality models for a given CV split setup)
compute_seasonality = False
apply_seasonality = False

# define train/test split date
if period == "3m":
    min_test_dt = "2015-03-01"
elif period == "1y":
    min_test_dt = "2019-10-01"
elif period == "":
    min_test_dt = "2019-01-01"
print(f"Min test set date for {period} dataset: {min_test_dt}")

# define what departure time variable is called
dep_utc_varname = "sched_depart_utc"

## Load data and perform simple transformations

# team_BASE_DIR = f"dbfs:/student-groups/Group_4_1"
# display(dbutils.fs.ls(f"{team_BASE_DIR}/interim/join_checkpoints"))

team_BASE_DIR = f"dbfs:/student-groups/Group_4_1"
# spark.sparkContext.setCheckpointDir(f"{team_BASE_DIR}/modeling_checkpoints")

# read in joined, cleaned dataset
# df = spark.read.parquet(f"{team_BASE_DIR}/interim/join_checkpoints/joined_flights_weather_{period}.parquet") # !!!
# df = spark.read.parquet(f"{team_BASE_DIR}/interim/join_checkpoints/joined_{period}_weather_cleaned_combo.parquet")
# df = spark.read.parquet(f"{team_BASE_DIR}/interim/join_checkpoints/joined_flights_weather{period}_v1.parquet")
# df = spark.read.parquet(f"{team_BASE_DIR}/interim/join_checkpoints/joined_{period}_timefeat.parquet")
# df = spark.read.parquet(f"{team_BASE_DIR}/interim/join_checkpoints/joined_{period}_timefeat_seasfeat.parquet")
df = spark.read.parquet(f"{team_BASE_DIR}/interim/join_checkpoints/joined_{period}_timefeat_seasfeat_cleaned.parquet")

# convert time variable to datetime
df = df.withColumn(dep_utc_varname, to_timestamp(col(dep_utc_varname)))

# add hour and date variables (needed for seasonality and CV splits, respectively)
df = df.withColumn("dep_hour_utc", f.hour(col(dep_utc_varname))) \
    .withColumn("dep_date_utc", to_date(col(dep_utc_varname)))

# define outcome variable
df = df.withColumn("outcome", (when((col("DEP_DELAY") >= 15) | (col("CANCELLED") == 1), 1).otherwise(0)).cast("double"))

# cast weather columns to double
weather_cols = [col for col in df.columns if "origin_Hourly" in col]
remove_me = ["origin_HourlyPresentWeatherType","origin_HourlySkyConditions","origin_HourlyWindDirection"]
num_weather_cols = [c for c in weather_cols if c not in remove_me]
for column in num_weather_cols:
    df = df.withColumn(column, col(column).cast("double"))

df.cache()


# split into train and test
df_train = df.filter(f.col(dep_utc_varname) < min_test_dt)
df_train.cache()
# print(f"Train data: {df_train.count()} records")
df_test = df.filter(f.col(dep_utc_varname) >= min_test_dt)
df_test.cache()
# print(f"Test data: {df_test.count()} records")

## Get cross-validation splits

# CODE IN THIS CELL DERIVED FROM DEMO 11 NOTEBOOK

def get_cv_time_limits_by_days_with_overlap(df, k=3, blocking=False, overlap=0, dep_utc_varname="dep_datetime", verbose=True):
    '''
    Get time bins for time-series cross validation, based on # days in dataset
    '''
    
    min_date = df.select(f.min("dep_date_utc")).collect()[0][0]
    max_date = df.select(f.max("dep_date_utc")).collect()[0][0]
    n_days = (max_date - min_date).days + 1
    total_width = k+1 - overlap*(k-1)
    chunk_size = np.ceil(n_days/total_width) # last chunk may be slightly smaller than the others

    # idx = np.arange(0,)
    # idx = np.arange(0,n_days,chunk_size)
    # idx[-1] = n_days-1
    # idx = [int(i)+1 for i in idx]
    
    if verbose:
        print(f'Splitting data into {k} folds with {overlap} overlap')
        print(f'Min date: {min_date}, max date: {max_date}')
        print(f'{chunk_size:,} days per fold')
        print("************************************************************")

    out = []
    for i in range(k):
        # define indices based on chunk size and overlap
        if i == 0:
            train_min_offset = 0
            train_max_offset = chunk_size
        else:
            train_min_offset += np.ceil((1-overlap)*chunk_size)
            train_max_offset += np.floor((1-overlap)*chunk_size)
        test_min_offset = train_max_offset
        test_max_offset = test_min_offset + chunk_size

        # define minimum training time based on cross-validation style
        if not blocking:
            t_min_train = min_date
        else:
            t_min_train = min_date + timedelta(days=train_min_offset)
        # define maximum training time
        t_max_train = min_date + timedelta(days=train_max_offset)
        # define minimum test time
        t_min_test = min_date + timedelta(days=test_min_offset)
        # define maximum test_time
        t_max_test = min_date + timedelta(days=test_max_offset)

        if t_max_test > max_date + timedelta(1):
            t_max_test = max_date + timedelta(1)

        out.append({"train_min":t_min_train, "train_max":t_max_train,
                    "test_min":t_min_test, "test_max":t_max_test})
    out = pd.DataFrame(out)
        
    if verbose:
        for i in range(k):
            print(f'    TRAIN set for fold {i} goes from {out["train_min"][i]} to {out["train_max"][i]}')
            print(f'    TEST set for fold {i} goes from {out["test_min"][i]} to {out["test_max"][i]}')
        print("(Note that the max dates are non-inclusive)")
        
    return out

# get cross-validation split times
cv_cutoffs = get_cv_time_limits_by_days_with_overlap(df_train.select("dep_date_utc"), k=k, blocking=True, overlap=overlap,
    dep_utc_varname=dep_utc_varname, verbose=True)
cv_cutoffs

## Train seasonality models for each fold

# informed by: https://www.databricks.com/blog/2021/04/06/fine-grained-time-series-forecasting-at-scale-with-facebook-prophet-and-apache-spark-updated-for-spark-3.html

def forecast_weekly_delay(history_pd: pd.DataFrame) -> pd.DataFrame:

    if history_pd[['ds','y']].dropna().shape[0] < 2:
        return pd.DataFrame({
            'dow': [0],
            'hour': [0],
            'weekly': [0],
            'daily': [0],
            'ORIGIN': [history_pd.ORIGIN.iloc[0]]
        })

    us_holidays = make_holidays_df(
        year_list=[2013 + i for i in range(10)], country='US'
    )
    
    # define Prophet model
    model = Prophet(
        interval_width=0.9,
        growth='linear',
        weekly_seasonality=True,
        daily_seasonality=True,
        yearly_seasonality=True,
        holidays=us_holidays,
        # seasonality_mode='multiplicative'
    )
    
    # fit the model
    model.fit(history_pd)
    
    # configure predictions for daily/weekly seasonality
    future_pd1 = model.make_future_dataframe(
        periods=24*7, 
        freq='h',
        include_history=False
    )
    
    # make predictions
    results_pd1 = model.predict(future_pd1)

    ### PREPARE DAILY/WEEKLY OUTPUT
    # ref date and dow
    ref_date = history_pd.ds.iloc[0].date()
    ref_dow = history_pd.DAY_OF_WEEK[0]
    # helper function: get day of the week,
    # using reference date and dow
    def get_dow(x,ref_date,dow):
        d_days = (x.date() - ref_date).days + dow
        d_days = d_days%7
        if d_days == 0:
            d_days = 7
        return d_days
    # get dow for forecasted points
    results_pd1['dow'] = results_pd1.ds.apply(lambda x: get_dow(x,ref_date,ref_dow))
    # get hour for forecasted points
    results_pd1['hour'] = results_pd1.ds.apply(lambda x: x.hour)
    # store origin
    results_pd1['ORIGIN'] = history_pd.ORIGIN.iloc[0]
    # get only needed columns
    results_pd1 = results_pd1[['dow','hour','weekly','daily','ORIGIN']]

    # return components
    return results_pd1

def forecast_annual_delay(history_pd: pd.DataFrame) -> pd.DataFrame:

    if history_pd[['ds','y']].dropna().shape[0] < 2:
        return pd.DataFrame({
            'month': [0],
            'dom': [0],
            'yearly': [0],
            'holidays': [0],  # Adjust this as needed
            'ORIGIN': [history_pd.ORIGIN.iloc[0]]
        })

    us_holidays = make_holidays_df(
        year_list=[2013 + i for i in range(10)], country='US'
    )
    
    # define Prophet model
    model = Prophet(
        interval_width=0.9,
        growth='linear',
        weekly_seasonality=True,
        daily_seasonality=True,
        yearly_seasonality=True,
        holidays=us_holidays,
        # seasonality_mode='multiplicative'
    )
    
    # fit the model
    model.fit(history_pd)
    
    # configure predictions for yearly/holiday seasonality
    future_pd2 = model.make_future_dataframe(
        periods=366, 
        freq='d',
        include_history=False
    )
    
    # make predictions
    results_pd2 = model.predict(future_pd2)

    ### PREPARE YEARLY/HOLIDAY OUTPUT
    # get month
    results_pd2['month'] = results_pd2.ds.apply(lambda x: x.month)
    # get day of month
    results_pd2['dom'] = results_pd2.ds.apply(lambda x: x.day)
    # store origin
    results_pd2['ORIGIN'] = history_pd.ORIGIN.iloc[0]
    # get only needed columns
    results_pd2 = results_pd2[['month','dom','yearly','holidays','ORIGIN']]
    has_leap_day = ((results_pd2.month == 2) & (results_pd2.dom == 29)).any()
    if not has_leap_day:
        results_pd2 = results_pd2.iloc[:-1, :]
        # estimate 2/29 result
        avg_yearly = results_pd2[
            ((results_pd2['month'] == 3) & (results_pd2['dom'] == 1)) |
            ((results_pd2['month'] == 2) & (results_pd2['dom'] == 28))
        ]['yearly'].mean()
        avg_holiday = results_pd2[
            ((results_pd2['month'] == 3) & (results_pd2['dom'] == 1)) |
            ((results_pd2['month'] == 2) & (results_pd2['dom'] == 28))
        ]['holidays'].mean()
        new_row = pd.DataFrame({
            'month': [2],
            'dom': [29],
            'yearly': [avg_yearly],
            'holidays': [0],  # Adjust this as needed
            'ORIGIN': [history_pd.ORIGIN.iloc[0]]
        })
        results_pd2 = pd.concat([results_pd2, new_row], ignore_index=True)

        
    # return components
    return results_pd2

schema1 = StructType([StructField('dow', LongType(), True),
                     StructField('hour', LongType(), True),
                     StructField('weekly', DoubleType(), True),
                     StructField('daily', DoubleType(), True),
                     StructField('ORIGIN', StringType(), True)])
schema2 = StructType([StructField('month', LongType(), True),
                     StructField('dom', LongType(), True),
                     StructField('yearly', DoubleType(), True),
                     StructField('holidays', DoubleType(), True),
                     StructField('ORIGIN', StringType(), True)])

def get_seasonality( df, t_min, t_max, forecast_fn, schema,
    dep_utc_varname=dep_utc_varname, delay_varname="DEP_DELAY"):

    input = df.filter((df[dep_utc_varname] >= t_min) & \
            (df[dep_utc_varname] < t_max)) \
            .withColumnRenamed(delay_varname, "y") \
            .withColumnRenamed(dep_utc_varname, "ds") \
            .groupBy('ORIGIN')
    result =  (input.applyInPandas(forecast_fn, schema=schema))
    
    return result

if compute_seasonality:

    # train seasonality model for each cross validation split
    for i in range(k):
        if i in [0,1,2]:
            print(f"Already calculated seasonality for fold {i+1}")
            continue

        print(f"training seasonality for CV fold {i + 1} of {k}")
        # train seasonality model for this cross validation split
        model_wd = get_seasonality(df_train, cv_cutoffs["train_min"][i], cv_cutoffs["train_max"][i],
                                   forecast_fn=forecast_weekly_delay, schema=schema1)
        model_yh = get_seasonality(df_train, cv_cutoffs["train_min"][i], cv_cutoffs["train_max"][i],
                                   forecast_fn=forecast_annual_delay, schema=schema2)
        # write out
        fn_wd = f"wd_seasonality_model_{period}_cv{i}of{k}_overlap{overlap}.parquet"
        model_wd.write.parquet(f"{team_BASE_DIR}/interim/{fn_wd}")
        fn_yh = f"yh_seasonality_model_{period}_cv{i}of{k}_overlap{overlap}.parquet"
        model_yh.write.parquet(f"{team_BASE_DIR}/interim/{fn_yh}")

    print(f"training seasonality for full training set")
    # train seasonality model for full training data
    model_wd = get_seasonality(df_train, datetime(1970,1,1), min_test_dt,
                             forecast_fn=forecast_weekly_delay, schema=schema1)
    model_yh = get_seasonality(df_train, datetime(1970,1,1), min_test_dt,
                             forecast_fn=forecast_annual_delay, schema=schema2)
    # write out
    fn_wd = f"wd_seasonality_model_{period}_train.parquet"
    model_wd.write.parquet(f"{team_BASE_DIR}/interim/{fn_wd}")
    fn_yh = f"yh_seasonality_model_{period}_train.parquet"
    model_yh.write.parquet(f"{team_BASE_DIR}/interim/{fn_yh}")

display(dbutils.fs.ls(f"{team_BASE_DIR}/interim/"))

### TODO: Rerun Seasonality Figures

airport = 'BOS'
seas1 = model_wd.filter(col('ORIGIN') == airport).toPandas()
seas2 = model_yh.filter(col('ORIGIN') == airport).toPandas()

fig,ax = plt.subplots(4,1,figsize=(10,6), gridspec_kw={'hspace': 0.6})
seas1['x'] = seas1['dow'] + seas1['hour']/24
seas1.sort_values('x',inplace=True)
seas1.plot(x='x',y='daily',ax=ax[0],legend=False)
seas1.plot(x='x',y='weekly',ax=ax[1],legend=False)
seas2.sort_values(['month','dom'],inplace=True)
seas2['x'] = range(1,seas2.shape[0]+1)
seas2.sort_values('x',inplace=True)
seas2.plot(x='x',y='yearly',ax=ax[2],legend=False)
seas2.plot(x='x',y='holidays',ax=ax[3],legend=False)
ax[0].set_title(f'{airport} Seasonality Components (minutes)\nTrained on 2015-2018 Data')
ax[0].set_xlabel('Day of week')
ax[1].set_xlabel('Day of week')
ax[0].set_ylabel('Daily')
ax[1].set_ylabel('Weekly')
ax[2].set_xlabel('Day of year')
ax[3].set_xlabel('Day of year')
ax[2].set_ylabel('Yearly')
ax[3].set_ylabel('Holiday')
plt.show()

## Apply seasonality models to full dataset

def get_seasonality_data(df, fold, k):
    """
    Look up seasonlaity features from saved seasonality model.
    """
    if fold == 'full':
        fn_wd = f"wd_seasonality_model_{period}_train.parquet"
        fn_yh = f"yh_seasonality_model_{period}_train.parquet"
    else:
        fn_wd = f"wd_seasonality_model_{period}_cv{fold}of{k}_overlap{overlap}.parquet"
        fn_yh = f"yh_seasonality_model_{period}_cv{fold}of{k}_overlap{overlap}.parquet"
    
    model_wd = spark.read.parquet(f"{team_BASE_DIR}/interim/{fn_wd}")
    model_yh = spark.read.parquet(f"{team_BASE_DIR}/interim/{fn_yh}")

    # add seasonality fields
    joined_df = df.alias("df") \
        .join(model_wd.alias("wd"), 
            (df["ORIGIN"] == model_wd["ORIGIN"]) & 
            (df["DAY_OF_WEEK"] == model_wd["dow"]) & 
            (df["dep_hour_utc"] == model_wd["hour"]),
            how="left") \
                .drop(model_wd["ORIGIN"]) \
                .drop(model_wd["dow"]) \
                .drop(model_wd["hour"]) \
        .join(model_yh.alias("yh"), 
            (df["ORIGIN"] == model_yh["ORIGIN"]) & 
            (df["MONTH"] == model_yh["month"]) & 
            (df["DAY_OF_MONTH"] == model_yh["dom"]),
            how="left") \
                .drop(model_yh["ORIGIN"]) \
                .drop(model_yh["month"]) \
                .drop(model_yh["dom"])
    
    # rename seasonality fields
    joined_df = joined_df \
        .withColumnRenamed("daily", f"daily_{fold}") \
        .withColumnRenamed("weekly", f"weekly_{fold}") \
        .withColumnRenamed("yearly", f"yearly_{fold}") \
        .withColumnRenamed("holidays", f"holidays_{fold}")

    return joined_df

if apply_seasonality:
    # apply CV features
    for i in range(k):
        df = get_seasonality_data(df, i, k)

    # apply full CV features
    df = get_seasonality_data(df, 'full', 5)


display(df)

fn_out = f"{team_BASE_DIR}/interim/join_checkpoints/joined_{period}_timefeat_seasfeat.parquet"
df.write.parquet(fn_out)

display(dbutils.fs.ls(f"{team_BASE_DIR}/interim/join_checkpoints"))

## Modeling

### Define columns to be used in model.

# weather columns
weather_cols = [col for col in df.columns if "origin_Hourly" in col]
remove_me = ["origin_HourlyPresentWeatherType","origin_HourlySkyConditions","origin_HourlyWindDirection"]
num_weather_cols = [c for c in weather_cols if c not in remove_me]

# seasonality columns
seasonality_cols = ["daily","weekly","yearly","holidays"]

# time columns
time_cols = ["mean_dep_delay","prop_delayed"]

# date related columns
date_cols = ["YEAR","MONTH","DAY_OF_MONTH","DAY_OF_WEEK"]

# flight metadata
flight_metadata_cols = ["OP_UNIQUE_CARRIER","ORIGIN_ICAO","DEST_ICAO"]

### TO ADD: PRIOR FLIGHT FEATURES, GRAPH FEATURES ###

# fields that will not be features but need to be kept for processing
keep_me = ["outcome","DAY_OF_WEEK","ORIGIN","dep_hour_utc",dep_utc_varname] # TODO: REVISIT WHICH ACTUALLY NEEDED

########## Define columns to be used as numeric and categorical features in the pipeline ##########
numeric_cols = [*num_weather_cols, *seasonality_cols, *time_cols]
categorical_cols = [*date_cols, *flight_metadata_cols]

### Define modeling pipeline

# List to hold the stages of the pipeline
stages = []

# Index and encode each categorical column
for column in categorical_cols:
    indexer = StringIndexer(inputCol=column, outputCol=column + "_index",handleInvalid="keep")
    encoder = OneHotEncoder(inputCol=column + "_index", outputCol=column + "_vec",handleInvalid="keep")
    stages += [indexer, encoder]
# define encoded categorical feature names
categorical_vec_columns = [col + "_vec" for col in categorical_cols]

# assemble features
features = numeric_cols + categorical_vec_columns
assembler = VectorAssembler(inputCols=features, outputCol="features", handleInvalid='keep')

# scale features
scaler = MinMaxScaler(inputCol="features", \
    outputCol="features_scaled")

# # logistic regression model
# lr = LogisticRegression(featuresCol='features_scaled', \
#     labelCol='outcome',maxIter=50)

xgb = SparkXGBClassifier(
  features_col="features_scaled",
  label_col="outcome"
)

# construct pipeline object from all components
pipeline = Pipeline(stages=stages+[assembler,scaler,xgb])

### Helper functions

# CODE BELOW DERIVED FROM DEMO 11 NOTEBOOK

def upsample(train_df,verbose=False):
  '''Upsamples train_df to balance classes'''
  #balance classes in train
  delay_count = train_df.filter(f.col("outcome") == 1).count()
  non_delay_count = train_df.filter(f.col("outcome") == 0).count()

  total = delay_count + non_delay_count
  keep_percent = non_delay_count / delay_count

  train_delay = train_df.filter(f.col('outcome') == 0)
  train_non_delay = train_df.filter(f.col('outcome') == 1).sample(withReplacement=True, fraction=keep_percent,seed=42)
  train_upsampled = train_delay.union(train_non_delay)
  return train_upsampled


def downsample(train_df,verbose=False):
  '''Downsamples train_df to balance classes'''
  #balance classes in train
  delay_count = train_df.filter(f.col("outcome") == 1).count()
  non_delay_count = train_df.filter(f.col("outcome") == 0).count()

  total = delay_count + non_delay_count
  keep_percent = delay_count / non_delay_count
  
  train_delay = train_df.filter(f.col('outcome') == 1)
  train_non_delay = train_df.filter(f.col('outcome') == 0).sample(withReplacement=False,fraction=keep_percent,seed=42)
  train_downsampled = train_delay.union(train_non_delay)
  return train_downsampled

def cv_eval(preds):
  """
  Input: transformed df with prediction and label
  Output: desired score 
  """
  rdd_preds_m = preds.select(['prediction', 'outcome']).rdd
  rdd_preds_b = preds.select('outcome','probability').rdd.map(lambda row: (float(row['probability'][1]), float(row['outcome'])))
  metrics_m = MulticlassMetrics(rdd_preds_m)
  metrics_b = BinaryClassificationMetrics(rdd_preds_b)
  F2 = np.round(metrics_m.fMeasure(label=1.0, beta=2.0), 4)
  pr = metrics_b.areaUnderPR
  return F2, pr

def timeSeriesSplitCV(df, pipeline, cv_info, sampling=None, metric='f2', verbose=True, dep_utc_varname=dep_utc_varname):
  '''
  Perform timSeriesSplit k-fold cross validation 
  '''

  k = len(cv_info)
  
  # Track score
  scores=[]
  
  # Start k-fold
  for i in range(k):
    
    # Create train set
    train_df = df.filter((df[dep_utc_varname] >= cv_info["train_min"][i]) & \
      (df[dep_utc_varname] < cv_info["train_max"][i])).cache()
      
    # Create dev set
    dev_df = df.filter((df[dep_utc_varname] >= cv_info["test_min"][i]) & \
      (df[dep_utc_varname] < cv_info["test_max"][i])).cache() 
    
    print(f"{train_df.count()} (unsampled) TRAIN records in fold {i+1}")
    print(f"{dev_df.count()} DEV records in fold {i+1}")

    # Apply sampling on train if selected
    if sampling=='down':
      train_df = downsample(train_df)
      train_df = train_df.cache()
    elif sampling=='up':
      train_df = upsample(train_df)
      train_df = train_df.cache()
    # elif sampling=='weights':
    #   train_df = add_class_weights(train_df).cache()
      
    # #print info on train and dev set for this fold
    # if verbose:
    #   print('    TRAIN set for fold {} goes from {} to {}, count is {:,} flights ({})'.format((i+1), 
    #                                                                                   train_df.agg({dep_utc_varname:'min'}).collect()[0][0],
    #                                                                                   train_df.agg({dep_utc_varname:'max'}).collect()[0][0],
    #                                                                                   train_df.count(),
    #                                                                                   sampling + '-sampled' if sampling else 'no sampling'))
    #   print('    DEV set for fold {} goes from {} to {}, count is {:,} flights'.format((i+1), 
    #                                                                                   dev_df.agg({dep_utc_varname:'min'}).collect()[0][0],
    #                                                                                   dev_df.agg({dep_utc_varname:'max'}).collect()[0][0],
    #                                                                                   dev_df.count()))
    
    
    # prep seasonality columns (rename, fill as needed)
    train_df = train_df \
      .withColumnRenamed(f"daily_{i}","daily") \
      .withColumnRenamed(f"weekly_{i}","weekly") \
      .withColumnRenamed(f"yearly_{i}","yearly") \
      .withColumnRenamed(f"holidays_{i}","holidays")
    train_df = train_df.fillna({col:0 for col in ['daily','weekly','yearly','holidays']})
    dev_df = dev_df \
      .withColumnRenamed(f"daily_{i}","daily") \
      .withColumnRenamed(f"weekly_{i}","weekly") \
      .withColumnRenamed(f"yearly_{i}","yearly") \
      .withColumnRenamed(f"holidays_{i}","holidays")
    dev_df = dev_df.fillna({col:0 for col in ['daily','weekly','yearly','holidays']})
        
    # Fit params on the model
    model = pipeline.fit(train_df)
    dev_pred = model.transform(dev_df)
    if metric=='f2':
      score = cv_eval(dev_pred)[0]
    elif metric=='pr':
      score = cv_eval(dev_pred)[1]
    scores.append(score)
    print(f'    Number of training datapoints for fold number {i+1} is {train_df.count():,} with a {metric} score of {score:.2f}') 
    print('------------------------------------------------------------')
  
  # Take average of all scores
  avg_score = np.average(scores)    
  print(f'Average {metric} score across all folds is {avg_score:.2f}')
  print("************************************************************")

  # # Train on full df
  # print('Training on full train dataset, and validating on dev dataset with best parameters from CV:')
  # print(best_parameters)
    
  # if verbose:
  #   print('    TRAIN set for best parameter fitted model goes from {} to {}, count is {:,} flights ({})'.format(train_df.agg({dep_utc_varname:'min'}).collect()[0][0],
  #                                                                                                    train_df.agg({dep_utc_varname:'max'}).collect()[0][0],
  #                                                                                                    train_df.count(),
  #                                                                                                    sampling + '-sampled' if sampling else 'no sampling'))
  return avg_score

### Train, cross validate, predict

timeSeriesSplitCV(df_train, pipeline, cv_cutoffs, sampling='down', metric='f2', verbose=True, dep_utc_varname=dep_utc_varname)

# final training and evaluation

# df_train = downsample(df_train).cache()
df_train = df_train \
    .withColumnRenamed(f"daily_full","daily") \
    .withColumnRenamed(f"weekly_full","weekly") \
    .withColumnRenamed(f"yearly_full","yearly") \
    .withColumnRenamed(f"holidays_full","holidays")
df_train = df_train.fillna({col:0 for col in ['daily','weekly','yearly','holidays']})
df_test = df_test \
    .withColumnRenamed(f"daily_full","daily") \
    .withColumnRenamed(f"weekly_full","weekly") \
    .withColumnRenamed(f"yearly_full","yearly") \
    .withColumnRenamed(f"holidays_full","holidays")
df_test = df_test.fillna({col:0 for col in ['daily','weekly','yearly','holidays']})
model = pipeline.fit(df_train)
dev_pred = model.transform(df_test)
# get f2 score
score = cv_eval(dev_pred)[0]
print(score)

## Plots for Report

outcome_cols = ['outcome']
outcome_info = df.select(outcome_cols).toPandas()

data = [np.sum(outcome_info['outcome']), np.sum(np.logical_not(outcome_info['outcome']))]
print(data)


data = [np.mean(outcome_info['outcome']), np.mean(np.logical_not(outcome_info['outcome']))]

fig,ax = plt.subplots(1,1,figsize=(10,2))
ax.barh([''],data[0], color='black',height=0.25,edgecolor='black',label='Delayed')
ax.barh([''],data[1], color='gray',height=0.25,edgecolor='black',left=data[0],label='Not Delayed')
ax.set_xlabel('Proportion')
ax.set_title('Outcome Variable Distribution')
ax.set_xlim((0,1))
ax.legend(loc='upper center',bbox_to_anchor=(0.5,-0.3),ncol=2)
plt.show()

# Plots for Report

fn_model = f"seasonality_model_{period}_train.parquet"
model = spark.read.parquet(f"{team_BASE_DIR}/interim/{fn_model}")
display(model)

airport = 'BOS'
seas = model.filter(col('ORIGIN') == airport).toPandas()

fig,ax = plt.subplots(2,1,figsize=(10,5))
seas['x'] = seas['dow'] + seas['hour']/24
seas.sort_values('x',inplace=True)
seas.plot(x='x',y='daily',ax=ax[0],legend=False)
seas.plot(x='x',y='weekly',ax=ax[1],legend=False)
ax[0].set_xticks([])
ax[0].set_xlabel('')
ax[0].set_title(f'{airport} Seasonality Components\nTrained on Jan-Sep 2019 Data')
ax[1].set_xlabel('Day of week')
ax[0].set_ylabel('Daily component (minutes)')
ax[1].set_ylabel('Weekly component (minutes)')
plt.show()

df = get_seasonality_data(df, 'full', 3).cache()

# Select the relevant columns and convert to Pandas DataFrame
selected_columns = ["daily", "weekly", "outcome"]
df_pd = df.select(selected_columns).toPandas()

# Calculate the correlation matrix
correlation_matrix = df_pd.corr(method='spearman')

# Plot the heatmap
plt.figure(figsize=(10, 8))
sns.heatmap(correlation_matrix, annot=True, cmap="coolwarm", vmin=-1, vmax=1)
plt.title("Spearman Correlation Heatmap\nSeasonality Features, 2019 (1 Year) Data")
plt.show()

## XGBoost
